requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
# created-by: conda 25.3.1
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
adjusttext=1.3.0=pypi_0
blas=1.0=mkl
bottleneck=1.3.7=py38ha9d4c09_0
brotli=1.0.9=h5eee18b_9
brotli-bin=1.0.9=h5eee18b_9
brotli-python=1.0.9=py38h6a678d5_8
bzip2=1.0.8=h5eee18b_6
c-ares=1.19.1=h5eee18b_0
ca-certificates=2025.2.25=h06a4308_0
certifi=2024.8.30=py38h06a4308_0
charset-normalizer=3.3.2=pyhd3eb1b0_0
contourpy=1.0.5=py38hdb19cb5_0
cuda-cudart=12.1.105=0
cuda-cupti=12.1.105=0
cuda-libraries=12.1.0=0
cuda-nvrtc=12.1.105=0
cuda-nvtx=12.1.105=0
cuda-opencl=12.9.19=0
cuda-runtime=12.1.0=0
cuda-version=12.9=3
cycler=0.11.0=pyhd3eb1b0_0
cyrus-sasl=2.1.28=h52b45da_1
dbus=1.13.18=hb2f20db_0
et-xmlfile=2.0.0=pypi_0
expat=2.7.1=h6a678d5_0
ffmpeg=4.3=hf484d3e_0
filelock=3.13.1=py38h06a4308_0
fontconfig=2.14.1=h55d465d_3
fonttools=4.51.0=py38h5eee18b_0
freetype=2.13.3=h4a9f257_0
fsspec=2025.3.0=pypi_0
giflib=5.2.2=h5eee18b_0
glib=2.78.4=h6a678d5_0
glib-tools=2.78.4=h6a678d5_0
gmp=6.3.0=h6a678d5_0
gmpy2=2.1.2=py38heeb90bb_0
gnutls=3.6.15=he1e5248_0
gst-plugins-base=1.14.1=h6a678d5_1
gstreamer=1.14.1=h5eee18b_1
icu=73.1=h6a678d5_0
idna=3.7=py38h06a4308_0
importlib-metadata=8.5.0=pypi_0
importlib_resources=6.4.0=py38h06a4308_0
intel-openmp=2023.1.0=hdb19cb5_46306
jinja2=3.1.4=py38h06a4308_0
joblib=1.4.2=pypi_0
jpeg=9e=h5eee18b_3
kiwisolver=1.4.4=py38h6a678d5_0
krb5=1.20.1=h143b758_1
lame=3.100=h7b6447c_0
lcms2=2.16=hb9589c4_0
ld_impl_linux-64=2.40=h12ee557_0
lerc=4.0.0=h6a678d5_0
libabseil=20250127.0=cxx17_h6a678d5_0
libbrotlicommon=1.0.9=h5eee18b_9
libbrotlidec=1.0.9=h5eee18b_9
libbrotlienc=1.0.9=h5eee18b_9
libclang=14.0.6=default_hc6dbbc7_2
libclang13=14.0.6=default_he11475f_2
libcublas=12.1.0.26=0
libcufft=11.0.2.4=0
libcufile=1.14.0.30=4
libcups=2.4.2=h2d74bed_1
libcurand=10.3.10.19=0
libcurl=8.12.1=hc9e6f67_0
libcusolver=11.4.4.55=0
libcusparse=12.0.2.55=0
libdeflate=1.22=h5eee18b_0
libedit=3.1.20230828=h5eee18b_0
libev=4.33=h7f8727e_1
libffi=3.4.4=h6a678d5_1
libgcc-ng=11.2.0=h1234567_1
libgfortran-ng=11.2.0=h00389a5_1
libgfortran5=11.2.0=h1234567_1
libglib=2.78.4=hdc74915_0
libgomp=11.2.0=h1234567_1
libiconv=1.16=h5eee18b_3
libidn2=2.3.4=h5eee18b_0
libjpeg-turbo=2.0.0=h9bf148f_0
libllvm14=14.0.6=hecde1de_4
libnghttp2=1.57.0=h2d74bed_0
libnpp=12.0.2.50=0
libnvjitlink=12.1.105=0
libnvjpeg=12.1.1.14=0
libpng=1.6.39=h5eee18b_0
libpq=17.4=hdbd6064_0
libprotobuf=5.29.3=hc99497a_0
libssh2=1.11.1=h251f7ec_0
libstdcxx-ng=11.2.0=h1234567_1
libtasn1=4.19.0=h5eee18b_0
libtiff=4.5.1=hffd6297_1
libunistring=0.9.10=h27cfd23_0
libuuid=1.41.5=h5eee18b_0
libwebp=1.3.2=h11a3e52_0
libwebp-base=1.3.2=h5eee18b_1
libxcb=1.17.0=h9b100fa_0
libxkbcommon=1.9.1=h69220b7_0
libxml2=2.13.8=hfdd30dd_0
llvm-openmp=14.0.6=h9e868ea_0
llvmlite=0.41.1=pypi_0
lz4-c=1.9.4=h6a678d5_1
markupsafe=2.1.3=py38h5eee18b_0
matplotlib=3.7.2=py38h06a4308_0
matplotlib-base=3.7.2=py38h1128e8f_0
minepy=1.2.6=pypi_0
mkl=2023.1.0=h213fc3f_46344
mkl-service=2.4.0=py38h5eee18b_1
mkl_fft=1.3.8=py38h5eee18b_0
mkl_random=1.2.4=py38hdb19cb5_0
mpc=1.3.1=h5eee18b_0
mpfr=4.2.1=h5eee18b_0
mpmath=1.3.0=py38h06a4308_0
mysql=8.4.0=h721767e_2
ncurses=6.4=h6a678d5_0
nettle=3.7.3=hbbd107a_1
networkx=3.1=py38h06a4308_0
numba=0.58.1=pypi_0
numexpr=2.8.4=py38hc78ab66_1
numpy=1.24.3=py38hf6e8229_1
numpy-base=1.24.3=py38h060ed82_1
nvidia-nccl-cu12=2.27.6=pypi_0
ocl-icd=2.3.2=h5eee18b_1
openh264=2.1.1=h4ff587b_0
openjpeg=2.5.2=he7f1fd0_0
openldap=2.6.4=h42fbc30_0
openpyxl=3.1.5=pypi_0
openssl=3.0.16=h5eee18b_0
packaging=24.1=py38h06a4308_0
pandas=2.0.3=py38h1128e8f_0
patsy=0.5.6=py38h06a4308_0
pcre2=10.42=hebb0a14_1
pillow=10.4.0=py38h5eee18b_0
pip=24.2=py38h06a4308_0
platformdirs=3.10.0=py38h06a4308_0
ply=3.11=py38_0
pooch=1.7.0=py38h06a4308_0
pthread-stubs=0.3=h0ce48e5_1
pynndescent=0.5.13=pypi_0
pyparsing=3.0.9=py38h06a4308_0
pyqt=5.15.10=py38h6a678d5_0
pyqt5-sip=12.13.0=py38h5eee18b_0
pysocks=1.7.1=py38h06a4308_0
python=3.8.20=he870216_0
python-dateutil=2.9.0post0=py38h06a4308_2
python-tzdata=2025.2=pyhd3eb1b0_0
pytorch=2.4.1=py3.8_cuda12.1_cudnn9.1.0_0
pytorch-cuda=12.1=ha16c6d3_6
pytorch-mutex=1.0=cuda
pytorch-tcn=1.2.3=pypi_0
pytz=2024.1=py38h06a4308_0
pyyaml=6.0.2=py38h5eee18b_0
qt-main=5.15.2=hb6262e9_12
readline=8.2=h5eee18b_0
requests=2.32.3=py38h06a4308_0
scikit-learn=1.3.2=pypi_0
scipy=1.10.1=py38hf6e8229_1
seaborn=0.13.2=pypi_0
setuptools=75.1.0=py38h06a4308_0
sip=6.7.12=py38h6a678d5_0
six=1.16.0=pyhd3eb1b0_1
sqlite=3.45.3=h5eee18b_0
statsmodels=0.14.0=py38ha9d4c09_0
sympy=1.13.3=py38h06a4308_0
tbb=2021.8.0=hdb19cb5_0
threadpoolctl=3.5.0=pypi_0
tk=8.6.14=h39e8969_0
tomli=2.0.1=py38h06a4308_0
torchaudio=2.4.1=py38_cu121
torchtriton=3.0.0=py38
torchvision=0.20.0=py38_cu121
tornado=6.4.1=py38h5eee18b_0
tqdm=4.66.5=py38h2f386ee_0
typing_extensions=4.11.0=py38h06a4308_0
umap=0.1.1=pypi_0
umap-learn=0.5.7=pypi_0
unicodedata2=15.1.0=py38h5eee18b_0
urllib3=2.2.3=py38h06a4308_0
wheel=0.44.0=py38h06a4308_0
xgboost=2.1.4=pypi_0
xkeyboard-config=2.44=h5eee18b_0
xlrd=2.0.2=pypi_0
xorg-libx11=1.8.12=h9b100fa_1
xorg-libxau=1.0.12=h9b100fa_0
xorg-libxdmcp=1.1.5=h9b100fa_0
xorg-xorgproto=2024.1=h5eee18b_1
xz=5.6.4=h5eee18b_1
yaml=0.2.5=h7b6447c_0
zipp=3.20.2=py38h06a4308_0
zlib=1.2.13=h5eee18b_1
zstd=1.5.6=hc292b87_0
